\name{RcppWavelet-package}
\alias{RcppWavelet-package}
\alias{RcppWavelet}
\docType{package}
\title{
\packageTitle{RcppWavelet}
}
\description{
\packageDescription{RcppWavelet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcppWavelet}
\packageIndices{RcppWavelet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RcppWavelet}

Maintainer: \packageMaintainer{RcppWavelet}
}
\references{
~~ Literature or other references for background information ~~
}
